Stop Poetry from setting Requires-Python: <4.0.0 (#22)
Using `^3.6` as per Poetry documentation is a bad idea for libraries. It
causes libraries to declare themselves incompatible with Python 4.0
which is silly.

Fixing this by setting the version constraint to `>=3.6`. To keep Poetry
happy I had to pin some dev dependencies to `^3.6` since the caret virus
has spread to other Python libraries. RIP Python 4.0.